Refactor: Move initialization logic to dedicated modules
Moves startup overlay, configuration, preferences, and field notes logic into their own modules to improve organization and maintainability.
